Fix backpack data not being loaded + more
- PropertiesBackpack gets added on entity init again
- initialized keeps track of if the entity had a backpack
- Added backpackHasItems packet, tells the client if a backpack has
items
- Prevented unequipping of backpacks that are opened by other players
- Refined backpack item description, depending if players access them

+++ src/net/mcft/copy/betterstorage/utils/PlayerUtils.java
@@ -24,7 +24,7 @@ public class PlayerUtils {
 		player.incrementWindowID();
 		
 		PacketUtils.sendPacket(player,
-				(byte)PacketHandler.openGui, player.currentWindowId,
+				PacketHandler.openGui, player.currentWindowId,
 				name, (byte)columns, (byte)rows, title);
 		
 		player.openContainer = container;
